Started rewrite for Tornado
This commit is my experiment with rewriting webvim using the tornado
module instead of pyramid. So far termio detects that a tornado.ioloop
is managing the web app and looks into the ioloop to give me async
callbacks on the `send_new_data` function so that we don't have to do
that weird while loop to detect updates.

This makes the responsiveness on the client side much faster as there is
no longer a .1 to 2 second delay depending on user activity in the
terminal.

Startup time has also been reduced since Tornado is a lighter weight web
framework than pyramid. This needs to be tested but just my initial
observations on this change over. So far a big win for the project.