Added `resize_check` method and terminal template.
Before this commit the terminal had strange behavior where when a resize
event would be called the terminal would revert back to 80x24 terminal
size. I have not had a chance to dig in to the code to really find out
why as it is a transient issue.

This commit adds in the `resize_check` method to make sure that the size
of the terminal matches that of the size that was previously requested.