Added proper text scaling
Before this commit the text scaling algorithm was based on us knowing
the width and height of the text that was being used instead of
calculating what the browser would render. This would result in
terminals in browsers that didn't render web fonts to be out of largely
out of scale.

This commit adds in the ability for you to use any font you wanted at
any size you wanted and the terminal would figure out how large the
terminal was in rows and columns.

The terminal also will be scaled to take up the entirety of the object
it is initialized on. This makes the terminal able to be embedded into
any page with the library in it.