Fallback to Base64 encoding only when needed
This commit adds in the functionality to automatically fallback to
base54 encoding and decoding for websocket payloads when the contents of
the payload cannot be encoded for send. In the case where a
`UnicodeDecodeError` is raised we base64 encode the string and send that
instead.

This will help reduce network bandwidth with the overhead of base64
encoding a string.